Fixes coveralls.io on pull requests
Coveralls.io now runs on pull requests as well.
The report is automatically uploaded on coveralls.io however it needs to
be accessed by copying the report URL to review it.
Next up would be to have it commenting on the pull request to see on
first sight if it increased or decreased.
Also made a couple of adjustment to the workflow file to clarify between
requirements building, Docker build/run and actual testing.